modified:   projects/server/src/controllers/authController1.js 	modified:   projects/server/src/controllers/userController1.js 	modified:   projects/server/src/index.js 	modified:   projects/server/src/middleware/auth.js 	deleted:    projects/server/src/migrations/20230907022211-create-user-voucher.js 	deleted:    projects/server/src/migrations/20230907040153-create-stock-history.js 	deleted:    projects/server/src/migrations/20230907041412-create-transaction-stock.js 	deleted:    projects/server/src/migrations/20230907053018-create-cart-stock.js 	modified:   projects/server/src/models/cart_stock.js 	modified:   projects/server/src/models/product.js 	modified:   projects/server/src/models/transaction.js 	modified:   projects/server/src/models/transaction_stock.js 	modified:   projects/server/src/models/user_voucher.js 	modified:   projects/server/src/router/index.js 	modified:   projects/server/src/router/userRouter.js 	new file:   projects/server/src/seeders/20230917091533-demo-voucher.js 	new file:   projects/server/src/seeders/20230917091820-demo-user-voucher.js

+++ projects/server/src/models/transaction.js
@@ -18,7 +18,7 @@ module.exports = (sequelize, DataTypes) => {
       id_user_address: DataTypes.INTEGER,
       totPrice: DataTypes.FLOAT,
       totQty: DataTypes.INTEGER,
-      totWeight: DataTypes.FLOAT,
+      totWeight: DataTypes.INTEGER,
       id_status: {
         type: DataTypes.INTEGER,
         defaultValue: 1,
